refactor: [#1720] extract singular-test handling into a dedicated RunTest class
Address review on #1879: singular-test classification and dispatch no
longer live inside `RunModel`. A new `RunTest` class owns the manifest
lookup and the `runTest(uri, testName)` call; the command registrations
in `src/commands/index.ts` consult `RunTest` first and delegate to
`RunModel` only when the active file is not a singular test.

This keeps `RunModel` model-only and removes its dependency on
`QueryManifestService`.

- New `src/commands/runTest.ts` with `RunTest`: `getSingularTestName`
  (manifest-driven classifier, honors project `test-paths` via the
  manifest) and `runSingularTestOnActiveWindowIfApplicable` (returns a
  boolean so callers can cleanly fall through to `RunModel`).
- Revert `RunModel` to its pre-#1720 shape (no `QueryManifestService`,
  no singular-test branches).
- Dispatch at the three affected command entry points:
  * `dbtPowerUser.runCurrentModel` — `dbt run` on a test file is never
    meaningful, route through `RunTest` first.
  * `dbtPowerUser.testCurrentModel` — singular tests must be selected
    by their own name, not by the surrounding model.
  * `dbtPowerUser.runTest` — tree-item invocation is unchanged; the
    command-palette (no tree item) path checks `RunTest` first.
- `src/inversify.config.ts` binds `RunTest` and injects it into
  `VSCodeCommands`.
- Tests moved from `runModel.test.ts` to `runTest.test.ts` and expanded
  to 8 cases covering the classifier + dispatch boolean contract.

Full Jest suite: 21 suites, 265 passed, 1 skipped, 0 failing.
`tsc --noEmit` clean.

Docker verification (`dbtpu-1720` container, jaffle-shop-duckdb +
`tests/singular_nonnull_orders.sql`): parsed the real `target/manifest.json`,
exercised the classifier against it, and confirmed the singular test
returns `"singular_nonnull_orders"` while a regular model returns
`undefined`. Verified `dist/extension.js` ships 4 call sites for
`runSingularTestOnActiveWindowIfApplicable` (the method definition +
the 3 command dispatchers) and no longer defines `getSingularTestName`
on `RunModel`.
