+++ src/inversify.config.ts
@@ -112,6 +112,7 @@ import { VSCodeCommands } from "./commands";
 import { AltimateScan } from "./commands/altimateScan";
 import { BigQueryCostEstimate } from "./commands/bigQueryCostEstimate";
 import { RunModel } from "./commands/runModel";
+import { RunTest } from "./commands/runTest";
 import { SqlToModel } from "./commands/sqlToModel";
 import { MissingSchemaTest } from "./commands/tests/missingSchemaTest";
 import { StaleModelColumnTest } from "./commands/tests/staleModelColumnTest";
@@ -1425,7 +1426,14 @@ container
 container
   .bind(RunModel)
   .toDynamicValue((context) => {
-    return new RunModel(
+    return new RunModel(context.container.get(DBTProjectContainer));
+  })
+  .inSingletonScope();
+
+container
+  .bind(RunTest)
+  .toDynamicValue((context) => {
+    return new RunTest(
       context.container.get(DBTProjectContainer),
       context.container.get(QueryManifestService),
     );
@@ -1533,6 +1541,7 @@ container
     return new VSCodeCommands(
       context.container.get(DBTProjectContainer),
       context.container.get(RunModel),
+      context.container.get(RunTest),
       context.container.get(SqlToModel),
       context.container.get(ValidateSql),
       context.container.get(AltimateScan),
